Don't build the jekyll site in migrations!

+++ engines/blog/lib/blog/engine.rb
@@ -5,24 +5,28 @@ module Blog
 
     isolate_namespace Blog
 
-    initializer "jekyll site" do |app|
-      jekylldir = Blog::Engine.root.join('jekyll').to_s
-      config = Blog::Engine.root.join('config', 'jekyll.yml').to_s
-      outroot = Blog::Engine.root.join('_site')
+    if defined?(::Rails::Server)
+      initializer "Jekyll" do |app|
+        Rails.logger = Logger.new(STDOUT)
 
-      FileUtils.rm_r outroot.to_s if File.exist? outroot.to_s
+        jekylldir = Blog::Engine.root.join('jekyll').to_s
+        config = Blog::Engine.root.join('config', 'jekyll.yml').to_s
+        outroot = Blog::Engine.root.join('_site')
 
-      Jekyll::Site.new(
-        Jekyll.configuration({
-          "config" => config,
-          "source" => jekylldir,
-          "destination" => outroot.join('ta').to_s
-        })
-      ).process
+        FileUtils.rm_r outroot.to_s if File.exist? outroot.to_s
 
-      puts "Built Jekyll Site!"
+        Jekyll::Site.new(
+          Jekyll.configuration({
+            "config" => config,
+            "source" => jekylldir,
+            "destination" => outroot.join('ta').to_s
+          })
+        ).process
 
-      app.middleware.use ::ActionDispatch::Static, outroot.to_s
+        Rails.logger.info("Jekyll site built!")
+
+        app.middleware.use ::ActionDispatch::Static, outroot.to_s
+      end
     end
   end
 end
